more modular code

+++ src/com/ledpixelart/console/GifDecoder.java
@@ -264,7 +264,7 @@ public class GifDecoder {
         int sx = i * iw; // start of line in source
         while (dx < dlim) {
           // map color and insert in destination
-          int index = ((int) pixels[sx++]) & 0xff;
+          int index = (pixels[sx++]) & 0xff;
           int c = act[index];
           if (c != 0) {
             dest[dx] = c;
@@ -430,7 +430,7 @@ public class GifDecoder {
               break;
             bi = 0;
           }
-          datum += (((int) block[bi]) & 0xff) << bits;
+          datum += ((block[bi]) & 0xff) << bits;
           bits += 8;
           bi++;
           count--;
@@ -470,7 +470,7 @@ public class GifDecoder {
           pixelStack[top++] = suffix[code];
           code = prefix[code];
         }
-        first = ((int) suffix[code]) & 0xff;
+        first = (suffix[code]) & 0xff;
 
         // Add a new string to the string table,
 
@@ -581,9 +581,9 @@ public class GifDecoder {
       int i = 0;
       int j = 0;
       while (i < ncolors) {
-        int r = ((int) c[j++]) & 0xff;
-        int g = ((int) c[j++]) & 0xff;
-        int b = ((int) c[j++]) & 0xff;
+        int r = (c[j++]) & 0xff;
+        int g = (c[j++]) & 0xff;
+        int b = (c[j++]) & 0xff;
         tab[i++] = 0xff000000 | (r << 16) | (g << 8) | b;
       }
     }
@@ -764,8 +764,8 @@ public class GifDecoder {
       readBlock();
       if (block[0] == 1) {
         // loop count sub-block
-        int b1 = ((int) block[1]) & 0xff;
-        int b2 = ((int) block[2]) & 0xff;
+        int b1 = (block[1]) & 0xff;
+        int b2 = (block[2]) & 0xff;
         loopCount = (b2 << 8) | b1;
       }
     } while ((blockSize > 0) && !err());
